+++ packages/fiori-app-sub-generator/src/fiori-app-generator/transforms.ts
@@ -101,6 +101,12 @@ export function transformTemplateType(
             };
         }
     }
+
+    // Add page title if addPageBuildingBlock is true
+    const pageBuildingBlockConfig = entityRelatedConfig?.addPageBuildingBlock
+        ? { pageBuildingBlockTitle: entityRelatedConfig.pageBuildingBlockTitle }
+        : {};
+
     const templateSettingsMap = {
         [TemplateTypeFE.ListReportObjectPage]: {
             entityConfig: _entityConfig,
@@ -132,6 +138,7 @@ export function transformTemplateType(
         },
         [TemplateTypeFE.FlexibleProgrammingModel]: {
             entityConfig: _entityConfig,
+            ...pageBuildingBlockConfig,
             pageName: FPM_DEFAULT_PAGE_NAME
         } as FPMSettings
     };
@@ -349,7 +356,8 @@ function getBaseAppConfig(
             addTests: canGenerateTests(template.type),
             generateIndex: generateIndexHtml,
             addAnnotations: entityRelatedConfig?.addFEOPAnnotations || entityRelatedConfig?.addLineItemAnnotations,
-            useVirtualPreviewEndpoints: project.enableVirtualEndpoints
+            useVirtualPreviewEndpoints: project.enableVirtualEndpoints,
+            addCdsUi5Plugin: project.addCdsUi5Plugin ?? true // Defaults to true
         },
         template: template as templateSetting extends BasicAppSettings
             ? TemplateSettingsFF<BasicAppSettings>