+++ packages/fiori-app-sub-generator/test/int/fiori-elements/expected-output/alp_v4_cap/app/alp_v4_cap/webapp/test/integration/pages/JourneyRunner.js
@@ -1,0 +1,19 @@
+sap.ui.define([
+    "sap/fe/test/JourneyRunner",
+	"testNameSpace/alpv4cap/test/integration/pages/BooksList",
+	"testNameSpace/alpv4cap/test/integration/pages/BooksObjectPage"
+], function (JourneyRunner, BooksList, BooksObjectPage) {
+    'use strict';
+
+    var runner = new JourneyRunner({
+        launchUrl: sap.ui.require.toUrl('testNameSpace/alpv4cap') + '/index.html',
+        pages: {
+			onTheBooksList: BooksList,
+			onTheBooksObjectPage: BooksObjectPage
+        },
+        async: true
+    });
+
+    return runner;
+});
+