+++ packages/generator-adp/src/app/index.ts
@@ -64,6 +64,7 @@ import { existsInWorkspace, showWorkspaceFolderWarning, handleWorkspaceFolderCho
 import { FDCService } from '@sap-ux/adp-tooling';
 import { getTargetEnvPrompt, getProjectPathPrompt } from './questions/target-env';
 import { isAppStudio } from '@sap-ux/btp-utils';
+import { getTemplatesOverwritePath } from '../utils/templates';
 
 const generatorTitle = 'Adaptation Project';
 
@@ -176,7 +177,7 @@ export default class extends Generator {
 
         if (!this.jsonInput) {
             this.env.lookup({
-                packagePatterns: ['@sap/generator-fiori', '@sap-ux/adp-flp-config-sub-generator']
+                packagePatterns: ['@sap/generator-fiori']
             });
             setHeaderTitle(opts, this.logger, generatorTitle);
 
@@ -266,15 +267,14 @@ export default class extends Generator {
             this._updateWizardStepsAfterNavigation();
 
             this.logger.info(`Project Attributes: ${JSON.stringify(this.attributeAnswers, null, 2)}`);
-
             if (this.attributeAnswers.addDeployConfig) {
-                const client = (await this.systemLookup.getSystemByName(this.configAnswers.system))?.Client;
+                const system = await this.systemLookup.getSystemByName(this.configAnswers.system);
                 addDeployGen(
                     {
                         projectName: this.attributeAnswers.projectName,
-                        targetFolder: this.attributeAnswers.targetFolder,
+                        projectPath: this.attributeAnswers.targetFolder,
                         connectedSystem: this.configAnswers.system,
-                        client
+                        system
                     },
                     this.composeWith.bind(this),
                     this.logger,
@@ -341,6 +341,10 @@ export default class extends Generator {
                 packageJson,
                 logger: this.toolsLogger
             });
+
+            if (config.options) {
+                config.options.templatePathOverwrite = getTemplatesOverwritePath();
+            }
 
             await generate(this._getProjectPath(), config, this.fs);
         } catch (e) {