+++ packages/odata-annotation-core-types/src/language-server.ts
@@ -3,6 +3,7 @@ export {
     Range,
     Position,
     Diagnostic,
+    DiagnosticTag,
     DiagnosticSeverity,
     TextEdit,
     WorkspaceEdit