+++ packages/ui5-application-inquirer/test/unit/prompts/prompt-helpers.test.ts
@@ -62,6 +62,13 @@ describe('prompt-helpers', () => {
     });
 
     test('hidePrompts', () => {
+        const mockCdsInfo = {
+            hasCdsUi5Plugin: true,
+            hasMinCdsVersion: true,
+            isCdsUi5PluginEnabled: true,
+            isWorkspaceEnabled: false
+        };
+
         const prompts: Record<Partial<promptNames>, UI5ApplicationQuestion> = {
             [promptNames.name]: {
                 name: promptNames.name
@@ -114,7 +121,7 @@ describe('prompt-helpers', () => {
         // All prompts returned
         expect(hidePrompts(prompts).length).toEqual(15);
         // Hide prompts that are not applicable for CAP projects
-        let filteredPrompts = hidePrompts(prompts, {}, true);
+        let filteredPrompts = hidePrompts(prompts, {}, mockCdsInfo);
         expect(filteredPrompts.length).toEqual(13);
         expect(filteredPrompts).not.toContainEqual({ name: promptNames.targetFolder });
         expect(filteredPrompts).not.toContainEqual({ name: promptNames.enableEslint });
@@ -148,17 +155,27 @@ describe('prompt-helpers', () => {
             }
         };
         // show `addDeployConfig` prompt when isCap is true
-        filteredPrompts = hidePrompts(prompts, promptOpts, true);
+        filteredPrompts = hidePrompts(prompts, promptOpts, mockCdsInfo);
         expect(filteredPrompts.length).toEqual(13);
         expect(filteredPrompts).toEqual(
             expect.arrayContaining([{ name: promptNames.addDeployConfig, when: expect.any(Function) }])
         );
 
         // hide `addDeployConfig` prompt when isCap is false
-        filteredPrompts = hidePrompts(prompts, promptOpts, false);
+        filteredPrompts = hidePrompts(prompts, promptOpts, undefined);
         expect(filteredPrompts.length).toEqual(14);
         expect(filteredPrompts).toEqual(
             expect.not.arrayContaining([{ name: promptNames.addDeployConfig, when: expect.any(Function) }])
+        );
+
+        // hide `enableTypeScript` and `enableVirtualEndpoints` when hasMinCdsVersion is false
+        filteredPrompts = hidePrompts(prompts, promptOpts, { ...mockCdsInfo, hasMinCdsVersion: false });
+        expect(filteredPrompts.length).toEqual(11);
+        expect(filteredPrompts).toEqual(
+            expect.not.arrayContaining([
+                { name: promptNames.enableTypeScript },
+                { name: promptNames.enableVirtualEndpoints }
+            ])
         );
     });
 });