+++ packages/cf-deploy-config-inquirer/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @sap-ux/cf-deploy-config-inquirer
 
+## 0.5.8
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.7.36
+
+## 0.5.7
+
+### Patch Changes
+
+-   178dbea: sanitize ignoreCertError (singular) configration option to ignoreCertErrors (plural)
+    -   @sap-ux/inquirer-common@0.7.35
+
+## 0.5.6
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.7.34
+
+## 0.5.5
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.7.33
+
+## 0.5.4
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.7.32
+
 ## 0.5.3
 
 ### Patch Changes