+++ packages/deploy-config-generator-shared/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @sap-ux/deploy-config-generator-shared
 
+## 0.1.29
+
+### Patch Changes
+
+-   @sap-ux/fiori-generator-shared@0.13.10
+
+## 0.1.28
+
+### Patch Changes
+
+-   @sap-ux/fiori-generator-shared@0.13.9
+
+## 0.1.27
+
+### Patch Changes
+
+-   @sap-ux/fiori-generator-shared@0.13.8
+
+## 0.1.26
+
+### Patch Changes
+
+-   @sap-ux/fiori-generator-shared@0.13.7
+
 ## 0.1.25
 
 ### Patch Changes