+++ packages/fiori-app-sub-generator/CHANGELOG.md
@@ -1,5 +1,173 @@
 # @sap-ux/fiori-app-sub-generator
 
+## 0.7.2
+
+### Patch Changes
+
+-   Updated dependencies [c6a0062]
+    -   @sap-ux/store@1.1.3
+    -   @sap-ux/odata-service-inquirer@2.6.1
+    -   @sap-ux/telemetry@0.6.17
+    -   @sap-ux/fiori-generator-shared@0.13.10
+    -   @sap-ux/cap-config-writer@0.12.1
+    -   @sap-ux/fiori-elements-writer@2.6.2
+    -   @sap-ux/fiori-freestyle-writer@2.4.38
+    -   @sap-ux/ui5-application-inquirer@0.15.2
+
+## 0.7.1
+
+### Patch Changes
+
+-   @sap-ux/fiori-elements-writer@2.6.1
+-   @sap-ux/fiori-freestyle-writer@2.4.37
+-   @sap-ux/odata-service-inquirer@2.6.0
+
+## 0.7.0
+
+### Minor Changes
+
+-   43f94eb: Add support for automatically including a Page Building Block when generating a custom page
+
+### Patch Changes
+
+-   Updated dependencies [43f94eb]
+    -   @sap-ux/odata-service-inquirer@2.6.0
+    -   @sap-ux/fiori-elements-writer@2.6.0
+
+## 0.6.24
+
+### Patch Changes
+
+-   Updated dependencies [8178b2b]
+    -   @sap-ux/cap-config-writer@0.12.0
+    -   @sap-ux/fiori-elements-writer@2.5.30
+    -   @sap-ux/fiori-freestyle-writer@2.4.36
+    -   @sap-ux/odata-service-inquirer@2.5.33
+    -   @sap-ux/ui5-application-inquirer@0.15.1
+
+## 0.6.23
+
+### Patch Changes
+
+-   Updated dependencies [810d7eb]
+    -   @sap-ux/annotation-generator@0.3.54
+    -   @sap-ux/fiori-elements-writer@2.5.29
+    -   @sap-ux/odata-service-inquirer@2.5.33
+
+## 0.6.22
+
+### Patch Changes
+
+-   @sap-ux/annotation-generator@0.3.53
+-   @sap-ux/fiori-elements-writer@2.5.28
+-   @sap-ux/odata-service-inquirer@2.5.33
+
+## 0.6.21
+
+### Patch Changes
+
+-   Updated dependencies [7772a11]
+    -   @sap-ux/ui5-application-inquirer@0.15.1
+
+## 0.6.20
+
+### Patch Changes
+
+-   55bb22e: updates cap projects to use npm workspaces by default
+-   Updated dependencies [55bb22e]
+    -   @sap-ux/ui5-application-inquirer@0.15.0
+    -   @sap-ux/cap-config-writer@0.11.0
+    -   @sap-ux/fiori-freestyle-writer@2.4.35
+    -   @sap-ux/fiori-elements-writer@2.5.27
+    -   @sap-ux/odata-service-inquirer@2.5.33
+
+## 0.6.19
+
+### Patch Changes
+
+-   178dbea: sanitize ignoreCertError (singular) configration option to ignoreCertErrors (plural)
+-   Updated dependencies [178dbea]
+    -   @sap-ux/ui5-application-inquirer@0.14.39
+    -   @sap-ux/fiori-freestyle-writer@2.4.34
+    -   @sap-ux/odata-service-inquirer@2.5.33
+    -   @sap-ux/fiori-elements-writer@2.5.26
+    -   @sap-ux/odata-service-writer@0.27.17
+    -   @sap-ux/launch-config@0.10.19
+    -   @sap-ux/project-access@1.30.12
+    -   @sap-ux/telemetry@0.6.16
+    -   @sap-ux/annotation-generator@0.3.52
+    -   @sap-ux/axios-extension@1.22.5
+    -   @sap-ux/cap-config-writer@0.10.27
+    -   @sap-ux/fiori-generator-shared@0.13.9
+
+## 0.6.18
+
+### Patch Changes
+
+-   @sap-ux/annotation-generator@0.3.51
+-   @sap-ux/fiori-elements-writer@2.5.25
+-   @sap-ux/project-access@1.30.11
+-   @sap-ux/odata-service-inquirer@2.5.32
+-   @sap-ux/axios-extension@1.22.5
+-   @sap-ux/cap-config-writer@0.10.26
+-   @sap-ux/fiori-generator-shared@0.13.8
+-   @sap-ux/launch-config@0.10.18
+-   @sap-ux/odata-service-writer@0.27.16
+-   @sap-ux/telemetry@0.6.15
+-   @sap-ux/ui5-application-inquirer@0.14.38
+-   @sap-ux/fiori-freestyle-writer@2.4.33
+
+## 0.6.17
+
+### Patch Changes
+
+-   @sap-ux/odata-service-inquirer@2.5.31
+-   @sap-ux/ui5-application-inquirer@0.14.37
+
+## 0.6.16
+
+### Patch Changes
+
+-   @sap-ux/fiori-elements-writer@2.5.24
+-   @sap-ux/odata-service-inquirer@2.5.30
+
+## 0.6.15
+
+### Patch Changes
+
+-   944af3e: Update all UI5 OPA test runner HTML files to use relative resource paths instead of absolute paths
+-   Updated dependencies [944af3e]
+    -   @sap-ux/fiori-elements-writer@2.5.23
+    -   @sap-ux/odata-service-inquirer@2.5.30
+    -   @sap-ux/fiori-freestyle-writer@2.4.32
+
+## 0.6.14
+
+### Patch Changes
+
+-   @sap-ux/fiori-elements-writer@2.5.22
+-   @sap-ux/odata-service-inquirer@2.5.30
+
+## 0.6.13
+
+### Patch Changes
+
+-   Updated dependencies [18a5ee2]
+    -   @sap-ux/telemetry@0.6.14
+    -   @sap-ux/fiori-generator-shared@0.13.7
+    -   @sap-ux/odata-service-inquirer@2.5.30
+    -   @sap-ux/cap-config-writer@0.10.25
+    -   @sap-ux/fiori-elements-writer@2.5.21
+    -   @sap-ux/fiori-freestyle-writer@2.4.31
+    -   @sap-ux/ui5-application-inquirer@0.14.36
+
+## 0.6.12
+
+### Patch Changes
+
+-   @sap-ux/fiori-elements-writer@2.5.20
+-   @sap-ux/odata-service-inquirer@2.5.29
+
 ## 0.6.11
 
 ### Patch Changes