+++ packages/preview-middleware-client/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @sap-ux-private/preview-middleware-client
 
+## 0.17.0
+
+### Minor Changes
+
+-   372e9ce: fix: [ADP][Info center] Remove all messages from the info center which overlap with messages added with the ui5 Log library.
+
+## 0.16.0
+
+### Minor Changes
+
+-   a39e0d9: fix: [ADP] For ui5 components which do not provide api.json we do not display Documentation error in the info center.
+
+## 0.15.3
+
+### Patch Changes
+
+-   ad49c30: fix: `Add Custom Table Column` Quick Action not using the correct fragment template.
+
 ## 0.15.2
 
 ### Patch Changes