+++ packages/project-integrity/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @sap-ux/project-integrity
 
+## 0.2.13
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.30.12
+
+## 0.2.12
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.30.11
+
+## 0.2.11
+
+### Patch Changes
+
+-   944af3e: Update all UI5 OPA test runner HTML files to use relative resource paths instead of absolute paths
+
 ## 0.2.10
 
 ### Patch Changes