+++ packages/ui-prompting/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @sap-ux/ui-prompting
 
+## 0.4.2
+
+### Patch Changes
+
+-   Updated dependencies [904f8e3]
+    -   @sap-ux/ui-components@1.26.10
+
+## 0.4.1
+
+### Patch Changes
+
+-   2d0380a: Improve answer change handling and dependent question choice requests in Questions component in ui prompting
+
+## 0.4.0
+
+### Minor Changes
+
+-   69cfbe3: add support for checked property in Select choices for default selection
+
+## 0.3.37
+
+### Patch Changes
+
+-   e4197e0: Support index-based default values for 'list' type questions. New property 'defaultIndex' for 'list' question.
+
 ## 0.3.36
 
 ### Patch Changes